backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytz==2023.3.post1
python-dotenv==1.0.0
secure-smtplib==0.1.1
sqlparse==0.4.4
tzdata==2023.4
gunicorn==21.2.0